Call clean_cached_assets randomly
